+++ src/components/common/TopNavigation.jsx
@@ -1,6 +1,6 @@
-import { useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
+import useAuthStore from '../../stores/useAuthStore';
 
 import LogoTextIcon from '../../assets/icons/NEXTVIBE.svg?react';
 import BadgeIcon from '../../assets/icons/badge.svg?react';
@@ -17,26 +17,32 @@ const NAV_ITEMS = [
 ];
 
 const TopNavigation = () => {
-  const [activePath, setActivePath] = useState('/home');
+  const navigate = useNavigate();
+  const location = useLocation();
+
+  // 전역 인증 상태
+  const isAuthenticated = useAuthStore((s) => s.isAuthenticated);
 
   return (
     <SHeaderWrap>
       <SBar>
-        <SBrand>
+        <SBrand
+          onClick={() => navigate('/home')}
+          role='button'
+          aria-label='홈으로 이동'
+        >
           <LogoIcon />
           <LogoTextIcon />
         </SBrand>
 
         <SNav>
           {NAV_ITEMS.map(({ path, label, Icon }) => {
-            const active = activePath.startsWith(path);
+            const active = location.pathname.startsWith(path);
             return (
               <SNavBtn
                 key={path}
                 to={path}
-                type='button'
                 $active={active}
-                onClick={() => setActivePath(path)}
                 aria-current={active ? 'page' : undefined}
               >
                 <SIcon $active={active}>
@@ -49,7 +55,25 @@ const TopNavigation = () => {
         </SNav>
 
         <SRight>
-          <UserIcon />
+          {isAuthenticated ? (
+            // 로그인 상태: 프로필 아이콘 (필요 시 드롭다운/마이페이지로 확장)
+            <ProfileBtn
+              type='button'
+              aria-label='프로필로 이동'
+              onClick={() => navigate('/mypage')}
+              title='프로필'
+            >
+              <UserIcon />
+            </ProfileBtn>
+          ) : (
+            // 로그아웃(미인증) 상태: 기존 before-login 우측 버튼 그대로
+            <>
+              <SButton onClick={() => navigate('/signup')}>회원가입</SButton>
+              <SButton onClick={() => navigate('/login')} $primary>
+                로그인
+              </SButton>
+            </>
+          )}
         </SRight>
       </SBar>
     </SHeaderWrap>
@@ -59,7 +83,6 @@ const TopNavigation = () => {
 export default TopNavigation;
 
 /* ---------------- styles ---------------- */
-
 const SHeaderWrap = styled.div`
   position: sticky;
   z-index: 1000;
@@ -84,6 +107,7 @@ const SBrand = styled.div`
   align-items: center;
   gap: 0.75rem;
   justify-self: start;
+  cursor: pointer;
   svg {
     height: 3.25rem;
     display: block;
@@ -118,11 +142,14 @@ const SNavBtn = styled(Link)`
     box-shadow: 0 0 0 3px rgba(81, 113, 255, 0.35);
   }
 `;
+
 const SRight = styled.div`
   display: flex;
   align-items: center;
+  gap: 10px;
   justify-self: end;
 `;
+
 const SIcon = styled.span`
   width: 1.75rem;
   height: 1.75rem;
@@ -145,4 +172,30 @@ const SLabel = styled.span`
   color: ${({ $active }) => ($active ? '#5C9DFF' : '#646879')};
   font-weight: ${({ $active }) => ($active ? 600 : 500)};
   transition: color 0.2s ease;
+`;
+
+/* before-login 우측 버튼 스타일 재사용 */
+const SButton = styled.button`
+  font-family: Pretendard;
+  font-size: 1.25rem;
+  font-style: normal;
+  font-weight: 600;
+  line-height: 1.25rem;
+
+  height: 2.75rem;
+  border-radius: 2.5rem;
+  padding: 8px 16px;
+  border: 1px solid ${({ $primary }) => ($primary ? '#DEEBFF' : '#ffffffff')};
+  background: ${({ $primary }) => ($primary ? '#DEEBFF' : 'transparent')};
+  color: ${({ $primary }) => ($primary ? '#646879' : 'var(--Gray-1, #646879)')};
+  cursor: pointer;
+  transition: all 0.2s ease;
+`;
+
+const ProfileBtn = styled.button`
+  background: transparent;
+  border: none;
+  padding: 0;
+  cursor: pointer;
+  line-height: 0;
 `;
