+++ src/stores/useAuthStore.js
@@ -1,0 +1,11 @@
+// src/stores/useAuthStore.js
+import { create } from 'zustand';
+
+const useAuthStore = create((set) => ({
+  user: null,
+  isAuthenticated: false,
+  login: (user) => set({ user, isAuthenticated: true }),
+  logout: () => set({ user: null, isAuthenticated: false }),
+}));
+
+export default useAuthStore;
